refactor(upload planner): Rename cost estimator to upload planner PE-458
