+++ src/arfs/arfs_upload_planner.test.ts
@@ -7,34 +7,34 @@ import {
 	stubPublicDriveMetaDataTx,
 	stubPublicFolderMetaDataTx
 } from '../../tests/stubs';
-import { ArFSTagSettings } from '../arfs/arfs_tag_settings';
+import { ArFSTagSettings } from './arfs_tag_settings';
 import { FeeMultiple, W } from '../types';
 import {
 	BundleRewardSettings,
 	CreateDriveV2TxRewardSettings,
 	EstimateCreateDriveParams
 } from '../types/cost_estimator_types';
-import { ArFSCostEstimator } from './arfs_cost_estimator';
-import { ARDataPriceChunkEstimator } from './ar_data_price_chunk_estimator';
-import { GatewayOracle } from './gateway_oracle';
+import { ArFSUploadPlanner } from './arfs_upload_planner';
+import { ARDataPriceChunkEstimator } from '../pricing/ar_data_price_chunk_estimator';
+import { GatewayOracle } from '../pricing/gateway_oracle';
 
-describe('The ArFSCostEstimator class', () => {
+describe('The ArFSUploadPlanner class', () => {
 	const arweaveOracle = new GatewayOracle();
 	const priceEstimator = new ARDataPriceChunkEstimator(true, arweaveOracle);
 	const arFSTagSettings = new ArFSTagSettings({ appName: 'Fabulous-Test', appVersion: '1.2' });
 
-	const bundledCostEstimator = new ArFSCostEstimator({
+	const bundledUploadPlanner = new ArFSUploadPlanner({
 		priceEstimator,
 		arFSTagSettings: arFSTagSettings
 	});
 
-	const v2TxCostEstimator = new ArFSCostEstimator({
+	const v2TxUploadPlanner = new ArFSUploadPlanner({
 		shouldBundle: false,
 		priceEstimator,
 		arFSTagSettings: arFSTagSettings
 	});
 
-	const boostedCostEstimator = new ArFSCostEstimator({
+	const boostedUploadPlanner = new ArFSUploadPlanner({
 		priceEstimator,
 		arFSTagSettings: arFSTagSettings,
 		feeMultiple: new FeeMultiple(10)
@@ -53,7 +53,7 @@ describe('The ArFSCostEstimator class', () => {
 			};
 
 			it('returns correct rewardSetting and totalWinstonPrice for a bundle', async () => {
-				const { rewardSettings, totalWinstonPrice } = await bundledCostEstimator.estimateCreateDrive(
+				const { rewardSettings, totalWinstonPrice } = await bundledUploadPlanner.estimateCreateDrive(
 					publicCreateDriveParams
 				);
 				const bundleRewardSettings = rewardSettings as BundleRewardSettings;
@@ -64,7 +64,7 @@ describe('The ArFSCostEstimator class', () => {
 			});
 
 			it('returns correct rewardSetting and totalWinstonPrice for a v2 transaction', async () => {
-				const { rewardSettings, totalWinstonPrice } = await v2TxCostEstimator.estimateCreateDrive(
+				const { rewardSettings, totalWinstonPrice } = await v2TxUploadPlanner.estimateCreateDrive(
 					publicCreateDriveParams
 				);
 				const v2RewardSettings = rewardSettings as CreateDriveV2TxRewardSettings;
@@ -77,7 +77,7 @@ describe('The ArFSCostEstimator class', () => {
 			});
 
 			it('returns correct rewardSetting and totalWinstonPrice for a fee boosted bundle', async () => {
-				const { rewardSettings, totalWinstonPrice } = await boostedCostEstimator.estimateCreateDrive(
+				const { rewardSettings, totalWinstonPrice } = await boostedUploadPlanner.estimateCreateDrive(
 					publicCreateDriveParams
 				);
 				const bundleRewardSettings = rewardSettings as BundleRewardSettings;
@@ -97,7 +97,7 @@ describe('The ArFSCostEstimator class', () => {
 			};
 
 			it('returns correct rewardSetting and totalWinstonPrice for a bundle', async () => {
-				const { rewardSettings, totalWinstonPrice } = await bundledCostEstimator.estimateCreateDrive(
+				const { rewardSettings, totalWinstonPrice } = await bundledUploadPlanner.estimateCreateDrive(
 					await privateCreateDriveParams()
 				);
 				const bundleRewardSettings = rewardSettings as BundleRewardSettings;
@@ -108,7 +108,7 @@ describe('The ArFSCostEstimator class', () => {
 			});
 
 			it('returns correct rewardSetting and totalWinstonPrice for a v2 transaction', async () => {
-				const { rewardSettings, totalWinstonPrice } = await v2TxCostEstimator.estimateCreateDrive(
+				const { rewardSettings, totalWinstonPrice } = await v2TxUploadPlanner.estimateCreateDrive(
 					await privateCreateDriveParams()
 				);
 				const v2RewardSettings = rewardSettings as CreateDriveV2TxRewardSettings;
